+++ src/ShapeCrawler/Colors/Color.cs
@@ -92,7 +92,6 @@ public struct Color
     ///     Creates color from Hex value.
     /// </summary>
     /// <param name="hex">Hex value.</param>
-    /// <param name="result">Color value.</param>
     /// <returns>Returns <see langword="true" /> if hex is a valid value. </returns>
     public static Color FromHex(string hex)
     {