+++ src/ShapeCrawler/ShapeCollection/SlideShapes.cs
@@ -129,14 +129,44 @@ internal sealed class SlideShapes : ISlideShapes
 
         if (skBitmap != null)
         {
+            var height = skBitmap.Height;
+            var width = skBitmap.Width;
+            var resize = false;
+
+            if (height > 500)
+            {
+                height = 500;
+                width = (int)(height * skBitmap.Width / (decimal)skBitmap.Height);
+                resize = true;
+            }
+
+            if (width > 500)
+            {
+                width = 500;
+                height = (int)(width * skBitmap.Height / (decimal)skBitmap.Width);
+                resize = true;
+            }
+
             var xEmu = UnitConverter.HorizontalPixelToEmu(100);
             var yEmu = UnitConverter.VerticalPixelToEmu(100);
-            var cxEmu = UnitConverter.HorizontalPixelToEmu(skBitmap.Width);
-            var cyEmu = UnitConverter.VerticalPixelToEmu(skBitmap.Height);
+            var cxEmu = UnitConverter.HorizontalPixelToEmu(width);
+            var cyEmu = UnitConverter.VerticalPixelToEmu(height);
 
-            var pPicture = this.CreatePPicture(imageStream, "Picture");
+            P.Picture? pPicture;
+            if (resize)
+            {
+                skBitmap.Resize(new SKSizeI(width:width, height:height), SKFilterQuality.High);
+                var resizedStream = new MemoryStream();
+                skBitmap.Encode(resizedStream, SKEncodedImageFormat.Png, 95);
+                resizedStream.Position = 0;
+                pPicture = this.CreatePPicture(resizedStream, "Picture");
+            }
+            else
+            {
+                pPicture = this.CreatePPicture(imageStream, "Picture");
+            }
 
-            var transform2D = pPicture.ShapeProperties!.Transform2D!;
+            var transform2D = pPicture!.ShapeProperties!.Transform2D!;
             transform2D.Offset!.X = xEmu;
             transform2D.Offset!.Y = yEmu;
             transform2D.Extents!.Cx = cxEmu;
@@ -599,6 +629,23 @@ internal sealed class SlideShapes : ISlideShapes
     {
         // Determine intrinsic size in 
         var size = GetSvgPixelSize(image);
+
+        // Ensure image size is not inserted at an unreasonable size
+        // See Issue #683 Large-dimension SVG files lead to error opening in PowerPoint
+        //
+        // Ideally, we'd want to use the slide dimensions itself. However, not sure how we get that
+        // here, so will use a fixed "safe" size
+        if (height > 500)
+        {
+            height = 500;
+            width = (int)(height * image.Width.Value / image.Height.Value);
+        }
+        
+        if (width > 500)
+        {
+            width = 500;
+            height = (int)(width * image.Height.Value / image.Width.Value);
+        }
 
         // Rasterize image at intrinsic size
         var bitmap = image.Draw((int)size.Width, (int)size.Height);