+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog  
 
+## Version 0.52.0 - Unreleased
+🍀Added support for the SVG format for the method `ISlideShapes.AddPicture()` [#350](https://github.com/ShapeCrawler/ShapeCrawler/issues/350)  
+🐞Fixed not displaying a picture added using the method `ISlideShapes.AddPicture()` [#649](https://github.com/ShapeCrawler/ShapeCrawler/issues/649)
+
 ## Version 0.51.0 - 2024-05-11
 🍀Added `IShapeFill.SetNoFill()` to remove shape filling [#667](https://github.com/ShapeCrawler/ShapeCrawler/issues/667)